Solucionar problema niveles #64
Se ha solucionado el problema que arrastraba el tema del nivel de los jugadores. Se ha llegado a la conclusión
del rango de niveles adecuado y se ha adaptado el código en consecuencia. closes #64

+++ pkg/modelos/grupo_test.go
@@ -18,14 +18,14 @@ func TestCrearJugador(t *testing.T) {
 	assert.Equal(t, true, success)
 	assert.Equal(t, 1, len(grupo.JugadoresNiveles))
 	assert.Equal(t, 1, len(grupo.JugadoresDisponibilidad))
-	assert.Equal(t, 50, grupo.JugadoresNiveles["Manuel"])
+	assert.Equal(t, uint(50), grupo.JugadoresNiveles["Manuel"])
 	assert.False(t, grupo.JugadoresDisponibilidad["Manuel"])
 
 	success, error = grupo.crearJugador("Jorge", 70, true)
 	assert.Equal(t, true, success)
 	assert.Equal(t, 2, len(grupo.JugadoresNiveles))
 	assert.Equal(t, 2, len(grupo.JugadoresDisponibilidad))
-	assert.Equal(t, 70, grupo.JugadoresNiveles["Jorge"])
+	assert.Equal(t, uint(70), grupo.JugadoresNiveles["Jorge"])
 	assert.True(t, grupo.JugadoresDisponibilidad["Jorge"])
 
 	success, error = grupo.crearJugador("Manuel", 75, false)
@@ -36,7 +36,7 @@ func TestCrearJugador(t *testing.T) {
 
 func TestCambiarDisponibilidadJugador(t *testing.T) {
 
-	JugadoresNiveles := map[string]int{"Manuel": 20}
+	JugadoresNiveles := map[string]uint{"Manuel": 20}
 	JugadoresDisponibilidad := map[string]bool{"Manuel": true}
 	grupo := Grupo{Nombre: "GrupoTest", JugadoresDisponibilidad: JugadoresDisponibilidad, JugadoresNiveles: JugadoresNiveles}
 
@@ -54,7 +54,7 @@ func TestCambiarDisponibilidadJugador(t *testing.T) {
 func TestCrearEquiposIgualadosParaPartido(t *testing.T) {
 
 	// Comprobamos que si no hay los suficientes jugadores disponibles no se pueden crear equipos igualados
-	JugadoresNiveles := map[string]int{"Manuel": 30, "Jorge": 50, "Edu": 10, "Clara": 90, "Migue": 100, "Alberto": 70,
+	JugadoresNiveles := map[string]uint{"Manuel": 30, "Jorge": 50, "Edu": 10, "Clara": 90, "Migue": 100, "Alberto": 70,
 		"Javi": 20, "Lorena": 80, "Maria": 60, "Sergio": 40}
 	JugadoresDisponibilidad := map[string]bool{"Manuel": false, "Jorge": false, "Edu": false, "Migue": true, "Clara": true, "Alberto": true,
 		"Javi": true, "Lorena": true, "Maria": true, "Sergio": false}
@@ -68,7 +68,7 @@ func TestCrearEquiposIgualadosParaPartido(t *testing.T) {
 
 	JugadoresDisponibilidad = map[string]bool{"Manuel": true, "Jorge": true, "Edu": true, "Migue": true, "Clara": true, "Alberto": true,
 		"Javi": true, "Lorena": true, "Maria": true, "Sergio": true}
-	JugadoresNivelesImposible := map[string]int{"Manuel": 3, "Jorge": 5, "Edu": 1, "Clara": 9, "Migue": 100, "Alberto": 7,
+	JugadoresNivelesImposible := map[string]uint{"Manuel": 3, "Jorge": 5, "Edu": 1, "Clara": 9, "Migue": 100, "Alberto": 7,
 		"Javi": 2, "Lorena": 8, "Maria": 6, "Sergio": 4}
 	grupo.JugadoresDisponibilidad = JugadoresDisponibilidad
 	grupo.JugadoresNiveles = JugadoresNivelesImposible
@@ -80,7 +80,7 @@ func TestCrearEquiposIgualadosParaPartido(t *testing.T) {
 
 	// Comprobamos caso correcto
 
-	JugadoresNiveles = map[string]int{"Manuel": 30, "Jorge": 50, "Edu": 10, "Clara": 90, "Migue": 100, "Alberto": 70,
+	JugadoresNiveles = map[string]uint{"Manuel": 30, "Jorge": 50, "Edu": 10, "Clara": 90, "Migue": 100, "Alberto": 70,
 		"Javi": 20, "Lorena": 80, "Maria": 60, "Sergio": 40}
 
 	grupo.JugadoresNiveles = JugadoresNiveles
